feat(auth): Next auth working

+++ src/actions/actions.ts
@@ -2,12 +2,94 @@
 
 import prisma from "@/lib/db";
 import { revalidatePath } from "next/cache";
-import { PetEssentials } from "@/lib/types";
-import { Pet } from "@prisma/client";
-import { petFormSchema, petIdSchema } from "@/lib/validation";
-import { getPetById } from "@/lib/server-utils";
+import { Prisma } from "@prisma/client";
+import { authSchema, petFormSchema, petIdSchema } from "@/lib/validation";
+import { checkAuth, getPetById } from "@/lib/server-utils";
+import { signIn, signOut } from "@/lib/auth-no-edge";
+import { AuthError } from "next-auth";
+import bcrypt from "bcryptjs";
+
+export async function logIn(prevState: unknown, formData: unknown) {
+  if (!(formData instanceof FormData)) {
+    return {
+      message: "Invalid form data.",
+    };
+  }
+
+  try {
+    await signIn("credentials", formData);
+  } catch (error) {
+    if (error instanceof AuthError) {
+      switch (error.type) {
+        case "CredentialsSignin": {
+          return {
+            message: "Invalid credentials.",
+          };
+        }
+        default: {
+          return {
+            message: "Error. Could not sign in.",
+          };
+        }
+      }
+    }
+
+    throw error; // nextjs redirects throws error, so we need to rethrow it
+  }
+}
+
+export async function signUp(prevState: unknown, formData: unknown) {
+  // check if formData is a FormData type
+  if (!(formData instanceof FormData)) {
+    return {
+      message: "Invalid form data.",
+    };
+  }
+
+  // convert formData to a plain object
+  const formDataEntries = Object.fromEntries(formData.entries());
+
+  // validation
+  const validatedFormData = authSchema.safeParse(formDataEntries);
+  if (!validatedFormData.success) {
+    return {
+      message: "Invalid form data.",
+    };
+  }
+
+  const { email, password } = validatedFormData.data;
+  const hashedPassword = await bcrypt.hash(password, 10);
+  try {
+    await prisma.user.create({
+      data: {
+        email,
+        hashedPassword,
+      },
+    });
+  } catch (error) {
+    if (error instanceof Prisma.PrismaClientKnownRequestError) {
+      if (error.code === "P2002") {
+        return {
+          message: "Email already exists.",
+        };
+      }
+    }
+
+    return {
+      message: "Could not create user.",
+    };
+  }
+
+  await signIn("credentials", formData);
+}
+
+export async function logOut() {
+  await signOut({ redirectTo: "/" });
+}
 
 export async function addPet(pet: unknown) {
+  const session = await checkAuth();
+
   const validatedPet = petFormSchema.safeParse(pet);
   if (!validatedPet.success) {
     return {
@@ -19,11 +101,11 @@ export async function addPet(pet: unknown) {
     await prisma.pet.create({
       data: {
         ...validatedPet.data,
-        // user: {
-        //   connect: {
-        //     id: session.user.id,
-        //   },
-        // },
+        user: {
+          connect: {
+            id: session.user.id,
+          },
+        },
       },
     });
   } catch (error) {
@@ -36,7 +118,11 @@ export async function addPet(pet: unknown) {
   revalidatePath("/app", "layout");
 }
 
-export async function editPet(petId: Pet["id"], newPetData: PetEssentials) {
+export async function editPet(petId: unknown, newPetData: unknown) {
+  // authentication check
+  const session = await checkAuth();
+
+  // validation
   const validatedPetId = petIdSchema.safeParse(petId);
   const validatedPet = petFormSchema.safeParse(newPetData);
 
@@ -46,14 +132,20 @@ export async function editPet(petId: Pet["id"], newPetData: PetEssentials) {
     };
   }
 
-  // Check if pet exists:
+  // authorization check
   const pet = await getPetById(validatedPetId.data);
   if (!pet) {
     return {
       message: "Pet not found.",
     };
   }
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "Not authorized.",
+    };
+  }
 
+  // database mutation
   try {
     await prisma.pet.update({
       where: {
@@ -71,6 +163,10 @@ export async function editPet(petId: Pet["id"], newPetData: PetEssentials) {
 }
 
 export async function deletePet(petId: unknown) {
+  // authentication check
+  const session = await checkAuth();
+
+  // validation
   const validatedPetId = petIdSchema.safeParse(petId);
   if (!validatedPetId.success) {
     return {
@@ -78,14 +174,20 @@ export async function deletePet(petId: unknown) {
     };
   }
 
-  // Check if pet exists:
+  // authorization check
   const pet = await getPetById(validatedPetId.data);
   if (!pet) {
     return {
       message: "Pet not found.",
     };
   }
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "Not authorized.",
+    };
+  }
 
+  // database mutation
   try {
     await prisma.pet.delete({
       where: {
